group: rename files
